ai-trading-agent/requirements.txt
# Core dependencies
ib_insync>=0.9.86
requests>=2.31
pandas>=2.0
pandas_ta>=0.3.14
pyyaml>=6.0
schedule>=1.2
python-dotenv>=1.0
tabulate>=0.9

# Dashboard
streamlit>=1.30
plotly>=5.18
streamlit-autorefresh>=1.0
pytz>=2023.3

# Telegram Bot
python-telegram-bot>=20.0

# AI Analysis
google-genai>=1.0

# Optional - advanced sentiment
# anthropic>=0.18
